Proposed/naïve fix for #8
This at least fixes the observed behavior in #8, but I am not sure if it
may cause some other undesirable side effect.

I had the idea for this fix in the course of testing out different
scenarios with the date input field. It turns out that if you edit the
field (and, as usual, experience the bug), and then hit the (x) button
to clear it, subsequent edits will not suffer from the strange behavior.
So it seemed to me that the solution was to provide the date input field
with an (empty) default value.

Does anyone have any insight as to whether this is an acceptable fix, or
if it will cause other problems?