Fix for #8
I had the idea for this fix in the course of testing out different
scenarios with the date input field. It turns out that if you edit the
field (and, as usual, experience the bug), and then hit the (x) button
to clear it, subsequent edits will not suffer from the strange behavior.
So it seemed to me that the solution was to provide the date input field
with an (empty) default value.