Use purescript-browserfeatures for feature detection
This is to (allow consumers of this library to) resolve #11. The
purescript-browserfeatures library provides a datatype to capture
support for the various values of the `type` attribute on `input`
elements; provided with a value of this type, `renderHalogen` will now
fall back gracefully to a `text` input type when an exotic one is
requested but not supported.

The purescript-browserfeatures library also provides detectors that use
the DOM to figure out what is supported by the browser; this is intended
to be used by clients of purescript-markdown-halogen in order to produce
the configuration that is to be passed to `renderHalogen`; that way,
this library is agnostic / open-ended as to the means of detecting
feature support.