Rewrite package along the general line of the new order!
The primary change here is to use purescript-markdown's `FormFieldP f`
type directly in the models, exploiting the fact that we may instantiate
the functor as we please.

I believe that these changes will allow us to address SD-1148 once and
for all.

It is possible and likely that there are newly introduced bugs in both
packages, as the changes were quite sweeping.